corrected options formatting and validations

+++ Infrastructure/DTOs/ExamDTOs/UpdateExamDTO.cs
@@ -10,7 +10,7 @@ namespace Infrastructure.DTOs.ExamDTOs
     {
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public decimal? TotalQuestions { get; set; }
+        public int? TotalQuestions { get; set; }
         public decimal? Duration { get; set; }
         public string? Tids { get; set; }
         public int? DisplayedQuestions { get; set; }
